EntWatch and Map Sync Fix
This will fix one bug related to time. If you press the button for the first time and wait for a full recharge. At a time when the time will approach 0, start spamming the E button, the plugin will think that everything went well, but the button on the map has not yet rolled back. Demonstrated this on video.